refactor(modules): restructure bot modules to mirror dashboard sidebar
Split admin.py → server_admin.py (modrole + language) + operator.py
(botpermissions + ping + prefix commands). Merge leavemsg.py into
moderation.py as /moderation leavemsg sub-group (breaking path change).
Merge rolemanage.py + reactionrole.py → roles.py regular Cog with two
explicit app_commands.Group class attributes. Update infrastructure
(auto_load, PROTECTED_MODULES, REQUIRED_PERMISSIONS, config template,
docker-compose), locale command-path values, all tests, and CLAUDE.md.
