+++ NerdyPy/config.yaml.template
@@ -17,15 +17,13 @@ bot:
   # description: "NerpyBot - Always one step ahead!"
   ops:
     - "your_discord_id_here"  # quoted to prevent formatter corruption
-  # admin and voicecontrol auto-load (admin always, voicecontrol with tagging/music)
+  # server_admin, operator, and voicecontrol auto-load (server_admin+operator always, voicecontrol with tagging/music)
   modules:
     - application
     - league
-    - leavemsg
     - moderation
-    - reactionrole
-    - rolemanage
     - reminder
+    - roles
     - tagging
     - wow
