+++ NerdyPy/locales/lang_de.yaml
@@ -27,7 +27,7 @@ admin:
     invalid: "Nicht unterstützte Sprache: `{language}`. Verfügbar: {available}"
   modrole:
     get_current: "Bot-Moderator-Rolle: **{role}**"
-    get_stale: "Eine Bot-Moderator-Rolle ist konfiguriert, aber die Rolle existiert nicht mehr. Verwende `/admin modrole delete` um sie zu entfernen."
+    get_stale: "Eine Bot-Moderator-Rolle ist konfiguriert, aber die Rolle existiert nicht mehr. Verwende `/modrole delete` um sie zu entfernen."
     get_none: "Keine Bot-Moderator-Rolle konfiguriert. Es werden berechtigungsbasierte Prüfungen verwendet."
     set_success: "Bot-Moderator-Rolle auf **{role}** gesetzt."
     delete_success: "Bot-Moderator-Rolle entfernt."
@@ -175,7 +175,7 @@ leavemsg:
   message:
     success: "Abschiedsnachricht aktualisiert auf: {message}"
     missing_placeholder: "Die Nachricht muss den Platzhalter {member} für den Mitgliedsnamen enthalten."
-    not_enabled: "Bitte aktiviere zuerst Abschiedsnachrichten mit `/leavemsg enable #channel`."
+    not_enabled: "Bitte aktiviere zuerst Abschiedsnachrichten mit `/moderation leavemsg enable #channel`."
     modal_title: "Abschiedsnachricht"
     modal_placeholder: "Verwende {member} für den Mitgliedsnamen, z.B. Tschüss {member}!"
   fetch_message:
