+++ NerdyPy/utils/permissions.py
@@ -23,20 +23,25 @@ REQUIRED_PERMISSIONS: dict[str, set[str]] = {
         "use_external_emojis",
         "use_external_stickers",
     },
-    "admin": set(),
+    "server_admin": set(),
+    "operator": set(),
     "application": {"send_messages", "embed_links", "create_public_threads"},
     "league": {"send_messages", "embed_links"},
-    "leavemsg": {"send_messages"},
-    "moderation": {"send_messages", "kick_members", "manage_messages", "manage_threads", "read_message_history"},
+    "moderation": {
+        "send_messages",
+        "kick_members",
+        "manage_messages",
+        "manage_threads",
+        "read_message_history",
+    },
     "music": {"send_messages", "embed_links", "connect", "speak"},
-    "reactionrole": {
+    "roles": {
         "send_messages",
         "manage_roles",
         "add_reactions",
         "manage_messages",
         "read_message_history",
     },
-    "rolemanage": {"send_messages", "manage_roles"},
     "reminder": {"send_messages"},
     "tagging": {"send_messages", "connect", "speak"},
     "voicecontrol": {"send_messages", "connect", "speak"},
