+++ tests/modules/test_roles_reactionrole.py
@@ -1,11 +1,11 @@
 # -*- coding: utf-8 -*-
-"""Tests for modules.reactionrole — localized reaction role responses."""
+"""Tests for modules.roles — reaction role commands."""
 
 from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 from models.reactionrole import ReactionRoleEntry, ReactionRoleMessage
-from modules.reactionrole import ReactionRole
+from modules.roles import Roles
 from utils.strings import load_strings
 
 
@@ -21,12 +21,12 @@ def _bypass_role_checks(monkeypatch):
     async def _always_true(*_args, **_kwargs):
         return True
 
-    monkeypatch.setattr("modules.reactionrole.is_role_below_bot", _always_true)
+    monkeypatch.setattr("modules.roles.is_role_below_bot", _always_true)
 
 
 @pytest.fixture
 def cog(mock_bot):
-    cog = ReactionRole.__new__(ReactionRole)
+    cog = Roles.__new__(Roles)
     cog.bot = mock_bot
     return cog
 
@@ -66,7 +66,7 @@ def role():
 
 class TestAdd:
     async def test_add_success(self, cog, interaction, channel, role):
-        await ReactionRole._add.callback(cog, interaction, channel, "12345", "👍", role)
+        await Roles._reactionrole_add.callback(cog, interaction, channel, "12345", "👍", role)
 
         msg = interaction.response.send_message.call_args[0][0]
         assert "Mapped" in msg
@@ -80,7 +80,7 @@ class TestAdd:
         db_session.add(ReactionRoleEntry(ReactionRoleMessageId=rr_msg.Id, Emoji="👍", RoleId=333))
         db_session.commit()
 
-        await ReactionRole._add.callback(cog, interaction, channel, "12345", "👍", role)
+        await Roles._reactionrole_add.callback(cog, interaction, channel, "12345", "👍", role)
 
         msg = interaction.response.send_message.call_args[0][0]
         assert "already mapped" in msg
@@ -93,7 +93,7 @@ class TestAdd:
 
 class TestRemove:
     async def test_remove_no_config(self, cog, interaction):
-        await ReactionRole._remove.callback(cog, interaction, "99999", "👍")
+        await Roles._reactionrole_remove.callback(cog, interaction, "99999", "👍")
 
         msg = interaction.response.send_message.call_args[0][0]
         assert "No reaction role config" in msg
@@ -105,10 +105,9 @@ class TestRemove:
         db_session.add(ReactionRoleEntry(ReactionRoleMessageId=rr_msg.Id, Emoji="👍", RoleId=333))
         db_session.commit()
 
-        # Mock _clear_reaction to avoid Discord API calls
         cog._clear_reaction = AsyncMock()
 
-        await ReactionRole._remove.callback(cog, interaction, "12345", "👍")
+        await Roles._reactionrole_remove.callback(cog, interaction, "12345", "👍")
 
         msg = interaction.response.send_message.call_args[0][0]
         assert "Removed mapping" in msg
@@ -121,7 +120,7 @@ class TestRemove:
 
 class TestList:
     async def test_list_empty(self, cog, interaction):
-        await ReactionRole._list.callback(cog, interaction)
+        await Roles._reactionrole_list.callback(cog, interaction)
 
         msg = interaction.response.send_message.call_args[0][0]
         assert "No reaction roles configured" in msg
@@ -134,7 +133,7 @@ class TestList:
 
 class TestClear:
     async def test_clear_no_config(self, cog, interaction):
-        await ReactionRole._clear.callback(cog, interaction, "99999")
+        await Roles._reactionrole_clear.callback(cog, interaction, "99999")
 
         msg = interaction.response.send_message.call_args[0][0]
         assert "No reaction role config" in msg
@@ -148,7 +147,7 @@ class TestClear:
 
         cog._clear_reaction = AsyncMock()
 
-        await ReactionRole._clear.callback(cog, interaction, "12345")
+        await Roles._reactionrole_clear.callback(cog, interaction, "12345")
 
         msg = interaction.response.send_message.call_args[0][0]
         assert "Cleared all" in msg
